src/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
appdirs==1.4.4
asgiref==3.7.2
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
boilerpy3==1.0.7
build==1.0.3
cachetools==5.3.2
camelot-py==0.11.0
cattrs==23.2.3
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.0
cryptography==42.0.2
cycler==0.12.1
dataclasses-json==0.6.3
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
effdet==0.4.1
emoji==2.10.0
et-xmlfile==1.1.0
Events==0.5
faiss-cpu==1.7.4
farm-haystack==1.24.0
fastapi==0.109.0
filelock==3.13.1
filetype==1.2.0
Flask==2.0.3
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
google-api-core==2.16.2
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
inflect==7.0.0
iopath==0.1.10
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.23
langchain-openai==0.0.5
langdetect==1.0.9
langsmith==0.0.87
layoutparser==0.3.4
lazy-imports==0.3.1
lxml==5.1.0
Markdown==3.5.2
MarkupSafe==2.1.4
marshmallow==3.20.2
matplotlib==3.8.2
mmh3==4.1.0
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
num2words==0.5.13
numpy==1.26.3
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==1.10.0
opencv-python==4.9.0.80
openpyxl==3.1.2
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.3
pikepdf==8.11.2
pillow==10.2.0
pinecone-client==3.0.2
platformdirs==4.2.0
poppler-utils==0.1.0
portalocker==2.8.2
posthog==3.3.4
prompthub-py==4.0.0
protobuf==4.25.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pulsar-client==3.4.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pydantic==1.10.14
pydantic_core==2.16.1
pymongo==4.6.2
pypandoc==1.12
pyparsing==3.1.1
pypdf==4.0.1
pypdfium2==4.26.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
python-iso639==2024.1.2
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.23
pytz==2023.4
PyYAML==6.0.1
quantulum3==0.9.0
rank-bm25==0.2.2
rapidfuzz==3.6.1
redis==5.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-cache==0.9.8
requests-oauthlib==1.3.1
rpds-py==0.17.1
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
sqlitedict==2.1.0
sseclient-py==1.8.0
starlette==0.35.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.2
timm==0.9.12
tokenizers==0.15.1
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.1
transformers==4.36.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
unstructured==0.12.3
unstructured-client==0.16.0
unstructured-inference==0.7.23
unstructured.pytesseract==0.3.12
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==2.2.0
uvicorn==0.27.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
Werkzeug==2.0.2
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.1.9
yarl==1.9.4
zipp==3.17.0
